구조체 별 파일 생성
